pyproject.toml
[tool.poetry]
name = "youtube-pinecone"
version = "0.1.0"
description = ""
authors = ["Sam <sammeehansam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
scrapetube = "^2.5.1"
pytube = "^15.0.0"
moviepy = "^1.0.3"
tqdm = "^4.66.1"
pydub = "^0.25.1"
pandas = "^2.2.0"
python-dotenv = "^1.0.1"
streamlit = "^1.30.0"
st-click-detector = "^0.1.3"
langchain = "^0.3.1"
cohere = "^4.45"
pinecone-client = "^5.0.0"
openai = "^1.10.0"
tiktoken = "^0.7.0"
yt-dlp = "^2024.8.6"
langchain-community = "^0.3.1"
langchain-openai = "^0.2.1"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
pydantic = "^2.9.2"
langchain-pinecone = "^0.2.0"
pyjwt = "^2.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
